+++ CHANGELOG.md
@@ -1,3 +1,23 @@
-This project is still active.
-Changes coming soon within the next few weeks.
-Stay tuned.
+Monday May 25, 2020
+
+### Removed:
+* Colormode removed (may be added again in future update).
+* No longer will display title information on every use.
+* -H flag changed to --lhost
+* -L flag changed to --lport
+* -S flag removed. Now language flags are used. (EX: --bash or --awk).
+* -l flag removed. Now --listener.
+* -d flag changed to --list to display supported languages.
+* -V flag changed to --version.
+* -s flag removed.
+
+### New:
+* Better command flags introduced.
+* --langs flag added.
+* --list flag added.
+* Edited help menu.
+* Install file introduced for installation.
+* Bash completeion file introduced.
+* New listener environment added (xfce).
+* Ability to save reverse shell automatically to clipboard. (Uses xclip).
+* Version information, title information, and contact information now moved under --version flag.